refactor: modularize product service, update config/infra, add observability docs and nginx conf
