+++ docker-compose.yml
@@ -3,39 +3,17 @@ networks:
     driver: bridge
 
 
-services:
-  product-service:
-    build:
-      context: .
-      dockerfile: ./product-service/Dockerfile
-    ports:
-      - "8082:8082"
-    volumes:
-      - ./product-service/data.json:/product-service/data.json
-    env_file:
-       - .env
-    environment:
-      - OTEL_ENDPOINT=otel-collector:4317
-      - PRODUCT_DATA_FILE_PATH=/product-service/data.json
-
-    networks:
-      - otel_internal-network
-    deploy:
-      replicas: 1
-      resources:
-        limits:
-          cpus: '0.5'
-          memory: 128M
-
-    
+services:    
   product-simulator:
     build:
       context: ./simulations
       dockerfile: Dockerfile
     environment:
-      - PRODUCT_SERVICE_URL=http://product-service:8082
+      - PRODUCT_SERVICE_URL=http://nginx:80
     networks:
       - otel_internal-network
+    depends_on:
+      - nginx
 
   otel-collector:
     image: otel/opentelemetry-collector-contrib:0.99.0
@@ -44,6 +22,7 @@ services:
     volumes:
       - ./otel-collector-config.yaml:/etc/otelcol-contrib/config.yaml:ro
       - /var/run/docker.sock:/var/run/docker.sock:ro
+      - /:/hostfs:ro
     ports:
       - "4317:4317"
       - "4318:4318"
@@ -55,6 +34,66 @@ services:
       - OTEL_RESOURCE_HOST_NAME=${OTEL_RESOURCE_HOST_NAME}
       - SIGNOZ_ENDPOINT=${SIGNOZ_ENDPOINT}
       - SIGNOZ_INGESTION_KEY=${SIGNOZ_INGESTION_KEY}
+
+
+
+  product-service-1:
+    build:
+      context: .
+      dockerfile: ./product-service/Dockerfile
+    volumes:
+      - ./product-service/data.json:/product-service/data.json
+    env_file:
+       - .env
+    environment:
+      - OTEL_ENDPOINT=otel-collector:4317
+      - PRODUCT_DATA_FILE_PATH=/product-service/data.json
+      - SERVICE_NAME=store_1
+      - SERVICE_VERSION=v1
+    networks:
+      - otel_internal-network
+    deploy:
+      replicas: 1
+      resources:
+        limits:
+          cpus: '0.5'
+          memory: 128M
+
+  
+  product-service-2:
+    build:
+      context: .
+      dockerfile: ./product-service/Dockerfile
+    volumes:
+      - ./product-service/data.json:/product-service/data.json
+    env_file:
+       - .env
+    environment:
+      - OTEL_ENDPOINT=otel-collector:4317
+      - PRODUCT_DATA_FILE_PATH=/data.json
+      - SERVICE_NAME=store_2
+      - SERVICE_VERSION=v2
+    networks:
+      - otel_internal-network
+    deploy:
+      replicas: 1
+      resources:
+        limits:
+          cpus: '0.5'
+          memory: 128M
+
+  nginx:
+    image: nginx:latest
+    ports:
+      - "8080:80"
+    volumes:
+      - ./nginx.conf:/etc/nginx/nginx.conf:ro
+    networks:
+      - otel_internal-network
+    depends_on:
+      - product-service-1
+      - product-service-2
+
