+++ nginx.conf
@@ -1,0 +1,28 @@
+events {}
+
+http {
+    # Define the group of backend servers
+    upstream product_backend {
+        # These are the internal service names and ports defined in docker-compose
+        server product-service-1:8082;
+        server product-service-2:8082;
+        # Nginx will load balance requests between these servers (default: round-robin)
+    }
+
+    server {
+        # Nginx listens on port 80 inside the container
+        listen 80;
+
+        # Handle requests to the root path
+        location / {
+            # Pass requests to the upstream group defined above
+            proxy_pass http://product_backend;
+            
+            # Recommended proxy headers
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header X-Forwarded-Proto $scheme;
+        }
+    }
+}
