+++ product-service/src/models/product_model.go
@@ -1,4 +1,4 @@
-package main
+package models
 
 const (
 	JSONFieldStock = "stock"
